requirements.txt
slackclient
boto3
flask